\name{teqr.DLdata}
\alias{teqr.DLdata}
\title{teqrOCtox intermediate file that creates dose level data}
\description{teqr.DLdata takes the simData output and creates the dose level data.
}
\usage{
teqr.DLdata(simData = simData)
}
\arguments{
  \item{simData}{includes the simulation results data,simresult, cohortSize,probt, MTDss, pTarget, lowerlimit, upperlimit,tootoxic, the number of simulation (sim)}
}
\references{Blanchard MS and Longmate JA, (2011). Toxicity equivalence range design (TEQR): A practical Phase I design. \emph{Contemporary Clinical Trials,(32)}, 114-121.}
\author{M.S. Blanchard <sblanchard@coh.org>}
\keyword{design}
